--- src/config/components/addRequirement/index.ts
@@ -1,20 +1,0 @@
-export const dataAddRequirement = {
-  title: "Agregar Requisito",
-  close: "Cerrar",
-  cancel: "Cancelar",
-  add: "Agregar",
-  labelPaymentMethod: "Tipo de requisito",
-  labelName: "Nombre del requisito",
-  labelAmount: "Cantidad",
-  labelValue: "Valor",
-  labelTextarea: "Descripción",
-  labelJustification: "Justificacion",
-  labelFrequency: "Frecuencia de pago",
-  labelDate: "Primer pago",
-  placeHolderSelect: "Seleccione una opción",
-  placeHolderAmount: "Número de pagos",
-  placeHolderValue: "Valor a pagar",
-  placeHolderDate: "Seleccione un requisito",
-  placeHolderTextarea: "Descripción del requisito",
-  placeHolderJustification: "Justificación del requisito",
-};
