+++ src/pages/board/outlets/financialReporting/Requirements/config.tsx
@@ -64,6 +64,28 @@ export const textFlagsRequirements = {
   titleError: "¡Uy, algo ha salido mal!",
   descriptionError: "No se han podido guardar los cambios.",
 };
+
+export const dataAddRequirement = {
+  title: "Agregar Requisito",
+  close: "Cerrar",
+  cancel: "Cancelar",
+  add: "Agregar",
+  labelPaymentMethod: "Tipo de requisito",
+  labelName: "Nombre del requisito",
+  labelAmount: "Cantidad",
+  labelValue: "Valor",
+  labelTextarea: "Descripción",
+  labelJustification: "Justificacion",
+  labelFrequency: "Frecuencia de pago",
+  labelDate: "Primer pago",
+  placeHolderSelect: "Seleccione una opción",
+  placeHolderAmount: "Número de pagos",
+  placeHolderValue: "Valor a pagar",
+  placeHolderDate: "Seleccione un requisito",
+  placeHolderTextarea: "Descripción del requisito",
+  placeHolderJustification: "Justificación del requisito",
+};
+
 export const infoItems = [
   { icon: <MdAddCircleOutline />, text: "Adjuntar", appearance: "help" },
   {
